requirements.txt
# Core
pydantic>=2.0
pyyaml>=6.0

# LLM
openai>=1.0
anthropic>=0.20.0

# Voice (optional - for local STT)
# sherpa-onnx>=1.0
# sounddevice>=0.4

# TTS
pyttsx3>=2.90
win10toast>=0.9

# System Control
pycaw>=20240216
pyautogui>=0.9.54
psutil>=5.9.0

# NLU Utilities (for ASR correction)
pypinyin>=0.50.0
rapidfuzz>=3.0.0
jieba>=0.42.1

# Utilities
python-dotenv>=1.0.0
